Renamed N* classes to no N prefix
