+++ src/main/java/net/jqwik/properties/stateful/RandomActionGenerator.java
@@ -5,13 +5,13 @@ import java.util.*;
 import net.jqwik.api.*;
 import net.jqwik.api.stateful.*;
 
-public class NRandomActionGenerator<T> implements NActionGenerator<T> {
+public class RandomActionGenerator<T> implements ActionGenerator<T> {
 
 	private final RandomGenerator<Action<T>> randomGenerator;
 	private final Random random;
 	private List<Shrinkable<Action<T>>> shrinkableActions = new ArrayList<>();
 
-	public NRandomActionGenerator(Arbitrary<Action<T>> actionArbitrary, int genSize, Random random) {
+	public RandomActionGenerator(Arbitrary<Action<T>> actionArbitrary, int genSize, Random random) {
 		this.random = random;
 		this.randomGenerator = actionArbitrary.generator(genSize);
 	}
