+++ src/main/java/net/jqwik/properties/stateful/SequentialActionSequence.java
@@ -5,107 +5,59 @@ import java.util.stream.*;
 
 import org.opentest4j.*;
 
-import net.jqwik.*;
-import net.jqwik.api.*;
 import net.jqwik.api.stateful.*;
 import net.jqwik.support.*;
 
 public class SequentialActionSequence<M> implements ActionSequence<M> {
 
-	public static <M> ActionSequenceArbitrary<M> fromActions(Arbitrary<Action<M>> actionArbitrary) {
-		return new DefaultActionSequenceArbitrary<>(actionArbitrary);
-	}
-
-	private final List<Action<M>> candidateSequence;
-	private final List<Try<M>> tries = new ArrayList<>();
+	private final ActionGenerator<M> actionGenerator;
+	private final int size;
+	private final List<Action<M>> sequence = new ArrayList<>();
 	private final List<Invariant<M>> invariants = new ArrayList<>();
 
-	private boolean hasRun = false;
+	private RunState runState = RunState.NOT_RUN;
 	private M state = null;
 
-	public SequentialActionSequence(List<Action<M>> candidateSequence) {
-		this.candidateSequence = candidateSequence;
+	public SequentialActionSequence(ActionGenerator<M> actionGenerator, int size) {
+		this.actionGenerator = actionGenerator;
+		this.size = size;
 	}
 
 	@Override
 	public synchronized List<Action<M>> runSequence() {
-		assertHasRun();
-		return tries //
-			.stream() //
-			.filter(Try::preconditionValid) //
-			.map(Try::action) //
-			.collect(Collectors.toList());
-	}
-
-	private void assertHasRun() {
-		if (!hasRun) {
-			throw new JqwikException("Sequence has not run yet.");
-		}
+		return sequence;
 	}
 
 	@Override
 	public synchronized M run(M model) {
-		tries.clear();
-		hasRun = true;
-		try {
-			state = tryAllCandidates(model);
+		if (runState != RunState.NOT_RUN) {
 			return state;
-		} catch (InvariantFailedError ife) {
-			throw ife;
-		} catch (Throwable t) {
-			AssertionFailedError assertionFailedError = new AssertionFailedError(createErrorMessage("Run", t.getMessage()), t);
-			assertionFailedError.setStackTrace(t.getStackTrace());
-			throw assertionFailedError;
 		}
-	}
-
-	@Override
-	public ActionSequence<M> withInvariant(Invariant<M> invariant) {
-		invariants.add(invariant);
-		return this;
-	}
-
-	@Override
-	public int size() {
-		return candidateSequence.size();
-	}
-
-	@Override
-	public M state() {
-		assertHasRun();
-		return state;
-	}
-
-	@Override
-	public String toString() {
-		String stateString;
-		List<Action<M>> actionsToShow;
-		if (hasRun) {
-			stateString = "(after run)";
-			actionsToShow = runSequence();
-		} else {
-			stateString = "(before run)";
-			actionsToShow = candidateSequence;
-		}
-		String actionsString = JqwikStringSupport.displayString(actionsToShow);
-		return String.format("%s %s:%s", this.getClass().getSimpleName(), stateString, actionsString);
-	}
-
-	private M tryAllCandidates(M model) {
+		runState = RunState.RUNNING;
 		state = model;
-		for (Action<M> candidate : candidateSequence) {
-			tryNextCandidate(candidate);
+		for (int i = 0; i < size; i++) {
+			Action<M> action;
+			try {
+				action = actionGenerator.next(state);
+			} catch (NoSuchElementException nsee) {
+				break;
+			}
+			sequence.add(action);
+			try {
+				state = action.run(state);
+				checkInvariants();
+			} catch (InvariantFailedError ife) {
+				runState = RunState.FAILED;
+				throw ife;
+			} catch (Throwable t) {
+				runState = RunState.FAILED;
+				AssertionFailedError assertionFailedError = new AssertionFailedError(createErrorMessage("Run", t.getMessage()), t);
+				assertionFailedError.setStackTrace(t.getStackTrace());
+				throw assertionFailedError;
+			}
 		}
+		runState = RunState.SUCCEEDED;
 		return state;
-	}
-
-	private void tryNextCandidate(Action<M> candidate) {
-		Try<M> aTry = new Try<>(candidate);
-		tries.add(aTry);
-		state = aTry.run(state);
-		if (aTry.preconditionValid()) {
-			checkInvariants();
-		}
 	}
 
 	private void checkInvariants() {
@@ -119,7 +71,7 @@ public class SequentialActionSequence<M> implements ActionSequence<M> {
 	}
 
 	private String createErrorMessage(String name, String causeMessage) {
-		String actionsString = tries
+		String actionsString = sequence
 			.stream() //
 			.map(aTry -> "    " + aTry.toString()) //
 			.collect(Collectors.joining(System.lineSeparator()));
@@ -132,36 +84,32 @@ public class SequentialActionSequence<M> implements ActionSequence<M> {
 		);
 	}
 
-	private static class Try<M> {
 
-		private final Action<M> action;
-		private boolean preconditionValid = false;
-
-		public Try(Action<M> action) {
-			this.action = action;
-		}
-
-		public M run(M model) {
-			if (action.precondition(model)) {
-				preconditionValid = true;
-				return action.run(model);
-			}
-			return model;
-		}
-
-		public boolean preconditionValid() {
-			return preconditionValid;
-		}
-
-		public Action<M> action() {
-			return action;
-		}
-
-		@Override
-		public String toString() {
-			String precondition = preconditionValid ? "" : " (precondition failed)";
-			return String.format("%s%s", action.toString(), precondition);
-		}
+	@Override
+	public synchronized ActionSequence<M> withInvariant(Invariant<M> invariant) {
+		invariants.add(invariant);
+		return this;
 	}
 
+	@Override
+	public int size() {
+		return size;
+	}
+
+	@Override
+	public RunState runState() {
+		return runState;
+	}
+
+	@Override
+	public synchronized M state() {
+		return state;
+	}
+
+	@Override
+	public String toString() {
+		String stateString = runState.name();
+		String actionsString = JqwikStringSupport.displayString(sequence);
+		return String.format("%s (%s) [%s]:%s", this.getClass().getSimpleName(), size, stateString, actionsString);
+	}
 }
